refactor: UserService 분리 (App & Transaction)
